+++ iExtra/Files/MainBundleImageFileFinder.swift
@@ -3,14 +3,13 @@
 //  iExtra
 //
 //  Created by Daniel Saidi on 2015-11-16.
-//  Copyright © 2015 Daniel Saidi. All rights reserved.
+//  Copyright © 2018 Daniel Saidi. All rights reserved.
 //
 
 /*
  
- This class will find all images in the app bundle and
- remove any retina variations. The class can only find
- images that are not placed in asset folders.
+ This class finds all non-asset images in the app bundle and
+ removes any retina variations.
  
  */
 
@@ -32,30 +31,28 @@ public class MainBundleImageFileFinder: MainBundleFileFinder {
         let result = filter(rawResult)
         return result
     }
+}
+
+
+// MARK: - Private Functions
+
+fileprivate extension MainBundleImageFileFinder {
     
+    func filter(_ rawResult: [String]) -> [String] {
+        return rawResult.flatMap { getBaseImageName($0) }
+    }
     
-    // MARK: Private methods
-    
-    private func filter(_ rawResult:[String]) -> [String] {
-        var result = [String]()
-        for fileName in rawResult {
-            if (isBaseImageName(fileName)) {
-                result.append(getBaseImageName(fileName))
-            }
-        }
+    func getBaseImageName(_ fileName: String) -> String? {
+        guard isBaseImageName(fileName) else { return nil }
+        var result = fileName.components(separatedBy: ".").first
+        result = result?.components(separatedBy: "~").first
         return result
     }
     
-    private func getBaseImageName(_ fileName:String) -> String {
-        var result = fileName.components(separatedBy: ".").first
-        result = result?.components(separatedBy: "~iphone").first
-        return result!
-    }
-    
-    private func isBaseImageName(_ fileName:String) -> Bool {
-        return fileName.range(of: "@") != nil
-            || fileName.range(of: "_2x") != nil
-            || fileName.range(of: "_3x") != nil
-            || fileName.range(of: "~ipad") != nil
+    func isBaseImageName(_ fileName: String) -> Bool {
+        return !fileName.contains("@")
+            && !fileName.contains("_2x")
+            && !fileName.contains("_3x")
+            && !fileName.contains("~")
     }
 }